feat: add Offer Info components for create offer
